Refactor Featurecraft components to use custom hook for managing assistant state and data fetching
