+++ avengers/frontend/webapp/src/app/assistants/featurecraft/components/assistant.js
@@ -1,70 +1,18 @@
-import { useState, useEffect } from "react";
-import axios from "axios";
 import styles from "@/app/page.module.css";
 import NewMessageBlock from "@/app/assistants/featurecraft/components/newMessageBlock";
 import MessageBlock from "@/app/assistants/featurecraft/components/messageBlock";
+import useFeaturecraftAssistant from "@/app/assistants/featurecraft/hooks/useFeaturecraftAssistant";
 
-export default function FeaturecraftAssistant({ conversationId, setConversationId }) {
-    const [members, setMembers] = useState(["You", "Gemini"]);
-    const [description, setDescription] = useState("Welcome to the Featurecraft Assistant!");
-    const [messages, setMessages] = useState([]);
-    const [pinnedMessages, setPinnedMessages] = useState([]);
-    const [totalMessages, setTotalMessages] = useState(0);
-
-    useEffect(() => {
-        async function fetchData() {
-            try {
-                const response = await axios.get(`http://localhost:8000/chat/${conversationId}`);
-                return response.data;
-            } catch (error) {
-                console.error("Error fetching chat:", error);
-            }
-        }
-
-        async function loadChatData() {
-            const chatData = await fetchData();
-            if (chatData) {
-                setMembers(chatData.members);
-                setDescription(chatData.description);
-                setMessages(chatData.messages);
-                setPinnedMessages(chatData.pinnedMessages);
-                setTotalMessages(chatData.totalMessages);
-            }
-        }
-
-        loadChatData();
-    }, [conversationId]);
-
-    const handleSendMessage = (sentMessage) => {
-        setMessages((prevMessages) => [
-            ...prevMessages,
-            sentMessage.newMessage
-        ]);
-    };
-
-    const handleReceiveMessage = (response) => {
-        console.log(response.data);
-
-        if (response.status === 200) {
-            setMessages((prevMessages) => [
-                ...prevMessages,
-                response.data
-            ]);
-        } else if (response.status === 201) {
-            setConversationId(response.data._id);
-            setMembers(response.data.members);
-            setDescription(response.data.description);
-            const assistantResponse = response.data.messages[response.data.messages.length - 1];
-            setMessages((prevMessages) => [
-                ...prevMessages,
-                assistantResponse
-            ]);
-            setPinnedMessages(response.data.pinnedMessages);
-            setTotalMessages(response.data.totalMessages);
-        } else {
-            console.error(response);
-        }
-    };
+export default function FeaturecraftAssistant({ conversationId, setConversationId, setAssistHistory }) {
+    const {
+        members,
+        description,
+        messages,
+        pinnedMessages,
+        totalMessages,
+        handleSendMessage,
+        handleReceiveMessage
+    } = useFeaturecraftAssistant(conversationId, setConversationId, setAssistHistory, true);
 
     return (
         <div className={`${styles.assistantInteraction} w-full h-full`}>
@@ -73,46 +21,19 @@ export default function FeaturecraftAssistant({ conversationId, setConversationI
                 <NewMessageBlock onSendMessage={handleSendMessage} onReceiveMessage={handleReceiveMessage} conversationId={conversationId} />
             </div>
         </div>
-    )
+    );
 }
 
-export function NewFeaturecraftAssistant({ conversationId, setConversationId }) {
-    const [members, setMembers] = useState(["You", "Gemini"]);
-    const [description, setDescription] = useState("Welcome to the Featurecraft Assistant!");
-    const [messages, setMessages] = useState([]);
-    const [pinnedMessages, setPinnedMessages] = useState([]);
-    const [totalMessages, setTotalMessages] = useState(0);
-
-    const handleSendMessage = (sentMessage) => {
-        setMessages((prevMessages) => [
-            ...prevMessages,
-            sentMessage.newMessage
-        ]);
-    };
-
-    const handleReceiveMessage = (response) => {
-        console.log(response.data);
-
-        if (response.status === 200) {
-            setMessages((prevMessages) => [
-                ...prevMessages,
-                response.data
-            ]);
-        } else if (response.status === 201) {
-            setConversationId(response.data._id);
-            setMembers(response.data.members);
-            setDescription(response.data.description);
-            const assistantResponse = response.data.messages[response.data.messages.length - 1];
-            setMessages((prevMessages) => [
-                ...prevMessages,
-                assistantResponse
-            ]);
-            setPinnedMessages(response.data.pinnedMessages);
-            setTotalMessages(response.data.totalMessages);
-        } else {
-            console.error(response);
-        }
-    };
+export function NewFeaturecraftAssistant({ conversationId, setConversationId, setAssistHistory }) {
+    const {
+        members,
+        description,
+        messages,
+        pinnedMessages,
+        totalMessages,
+        handleSendMessage,
+        handleReceiveMessage
+    } = useFeaturecraftAssistant(conversationId, setConversationId, setAssistHistory, false);
 
     return (
         <div className={`${styles.assistantInteraction} w-full h-full`}>
@@ -121,5 +42,5 @@ export function NewFeaturecraftAssistant({ conversationId, setConversationId }) 
                 <NewMessageBlock onSendMessage={handleSendMessage} onReceiveMessage={handleReceiveMessage} conversationId={conversationId} />
             </div>
         </div>
-    )
+    );
 }
