+++ avengers/frontend/webapp/src/app/assistants/featurecraft/page.js
@@ -9,7 +9,7 @@ import '@/app/globals.css';
 
 import AssistantPicker from "@/app/components/assistantPicker";
 import AssistantHistory from "@/app/components/assistantHistory";
-import {NewFeaturecraftAssistant} from "@/app/assistants/featurecraft/components/assistant";
+import { NewFeaturecraftAssistant } from "@/app/assistants/featurecraft/components/assistant";
 
 export default function Interactor() {
     const assistName = "FeatureCraft";
@@ -17,7 +17,6 @@ export default function Interactor() {
     const [assistHistory, setAssistHistory] = useState([]);
     const [conversationId, setConversationId] = useState("673d10aa792f5dc2123f0895"); // TODO: Change backend to accept "new" instead of an ID
 
-    
     useEffect(() => {
         axios.get("http://localhost:8000/history")
             .then(response => {
@@ -39,12 +38,11 @@ export default function Interactor() {
             .catch(error => console.error("Error fetching history:", error));
     }, []);
 
-
     return (
         <div className={styles.interactorLayout}>
             <AssistantPicker />
-            <AssistantHistory name={assistName} type={assistType} interactions={assistHistory}/>
-            <NewFeaturecraftAssistant conversationId={conversationId} setConversationId={setConversationId} />
+            <AssistantHistory name={assistName} type={assistType} interactions={assistHistory} />
+            <NewFeaturecraftAssistant conversationId={conversationId} setConversationId={setConversationId} setAssistHistory={setAssistHistory} />
         </div>
-    )
+    );
 }
